unilink/backend/requirements.txt
annotated-types==0.6.0
anyio==4.3.0
bcrypt==4.0.1
cassandra-driver==3.29.1
cffi==1.16.0
click==8.1.7
colorama==0.4.6
cryptography==42.0.5
dnspython==2.6.1
ecdsa==0.19.0
email_validator==2.1.1
fastapi==0.110.1
geomet==0.2.1.post1
h11==0.14.0
httptools==0.6.1
idna==3.7
kafka-python==2.0.2
passlib==1.7.4
pyasn1==0.6.0
pycparser==2.22
pydantic==2.6.4
pydantic_core==2.16.3
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.9
PyYAML==6.0.1
redis==5.0.3
rsa==4.9
six==1.16.0
sniffio==1.3.1
starlette==0.37.2
typing_extensions==4.11.0
uvicorn==0.29.0
watchfiles==0.21.0
websockets==12.0
